requirements.txt
flask==1.1.2
python-dotenv==0.14.0
PyYaml==5.3.1
requests==2.24.0
